+++ src/application-insights/HISTORY.rst
@@ -1,3 +1,18 @@
+0.1.12
+++++++++++++++++++
+
+* `az monitor app-insights component connect-webapp`: Enable application insights on web app.
+
+0.1.11
+++++++++++++++++++
+
+* `az monitor app-insights component update`: Update consent message.
+
+0.1.10
+++++++++++++++++++
+
+* `az monitor app-insights component update`: Prompt consent when migrating to workspace-centric workspace.
+
 0.1.9
 ++++++++++++++++++