+++ src/image-copy/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.2.8
+++++++
+* Remove unused --subscription parameter
+
+0.2.7
+++++++
+* Fix copying failure when location of source resource group and source image are different.
+
 0.2.6
 ++++++
 * Add validation on temporary resource group.